DEF-1604: Courses being correctly pulled.
