+++ autoconfigaicc.php
@@ -18,8 +18,9 @@
  * AICC auto configuration page
  *
  * @package    tool_opensesame
- * @subpackage opensesame
- * @copyright  2023 Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @copyright  2023 Moodle
+ * @author     Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @author     David Castro <david.castro@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
