+++ classes/auto_config.php
@@ -17,7 +17,9 @@
 /**
  * AICC auto configuration tool.
  * @package     tool_opensesame
- * @copyright   2023 Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @copyright   2023 Moodle
+ * @author      Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @author      David Castro <david.castro@moodle.com>
  * @license     https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
