+++ classes/task/opensesamesync.php
@@ -15,9 +15,7 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 namespace tool_opensesame\task;
 
-use context_course;
-use tool_opensesame\api\opensesame;
-use tool_opensesame\opensesame_handler;
+use tool_opensesame\local\opensesame_handler;
 
 defined('MOODLE_INTERNAL') || die();
 require_once($CFG->dirroot . '/lib/filelib.php');
@@ -29,7 +27,9 @@ require_once($CFG->dirroot . '/course/lib.php');
  *
  * @since      3.9
  * @package    tool_opensesame
- * @copyright  2023 Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @copyright  2023 Moodle
+ * @author     Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @author     David Castro <david.castro@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class opensesamesync extends \core\task\scheduled_task {
@@ -52,10 +52,9 @@ class opensesamesync extends \core\task\scheduled_task {
      * @throws \dml_exception
      */
     public function execute(): bool {
-
         mtrace("Opensesame task just started.");
         $handler = new opensesame_handler();
-        $handler->retrieve_courses();
+        $handler->run();
         mtrace('opensesame just finished.');
         return true;
     }
