+++ lang/en/tool_opensesame.php
@@ -18,7 +18,9 @@
  * Strings for component 'tool_opensesame', language 'en'
  *
  * @package    tool_opensesame
- * @copyright  2023 Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @copyright  2023 Moodle
+ * @author Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @author David Castro <david.castro@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
