+++ settings.php
@@ -18,8 +18,9 @@
  * Capability overview settings
  *
  * @package    tool_opensesame
- * @subpackage opensesame
- * @copyright  2023 Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @copyright  2023 Moodle
+ * @author     Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @author     David Castro <david.castro@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
@@ -36,18 +37,16 @@ if ($hassiteconfig) {
     $settings = new admin_settingpage('tool_opensesame', get_string('opensesameintegration', 'tool_opensesame'));
     // Creating new settings to add the new settingspage.
 
-    if (!is_array($CFG->forced_plugin_settings['tool_opensesame'])) {
-        $settings->add(new admin_setting_configtext('tool_opensesame/clientid', get_string('clientid', 'tool_opensesame'),
-                get_string('clientiddesc', 'tool_opensesame'), '', PARAM_RAW));
-        $settings->add(new admin_setting_configpasswordunmask('tool_opensesame/clientsecret',
-                get_string('clientsecret', 'tool_opensesame'),
-                get_string('clientsecretdesc', 'tool_opensesame'), ''));
-        $settings->add(new admin_setting_configtext('tool_opensesame/authurl', get_string('authurl', 'tool_opensesame'),
-                get_string('authurldesc', 'tool_opensesame'), 'https://auth.coursecloud.net/oauth2/aus1l01v8s55riV0C0h8/v1/token',
-                PARAM_URL));
-        $settings->add(new admin_setting_configtext('tool_opensesame/baseurl', get_string('baseurl', 'tool_opensesame'),
-                get_string('baseurldesc', 'tool_opensesame'), 'https://api.delivery.opensesame.com', PARAM_URL));
-    }
+    $settings->add(new admin_setting_configtext('tool_opensesame/clientid', get_string('clientid', 'tool_opensesame'),
+            get_string('clientiddesc', 'tool_opensesame'), '', PARAM_RAW));
+    $settings->add(new admin_setting_configpasswordunmask('tool_opensesame/clientsecret',
+            get_string('clientsecret', 'tool_opensesame'),
+            get_string('clientsecretdesc', 'tool_opensesame'), ''));
+    $settings->add(new admin_setting_configtext('tool_opensesame/authurl', get_string('authurl', 'tool_opensesame'),
+            get_string('authurldesc', 'tool_opensesame'), 'https://auth.coursecloud.net/oauth2/aus1l01v8s55riV0C0h8/v1/token',
+            PARAM_URL));
+    $settings->add(new admin_setting_configtext('tool_opensesame/baseurl', get_string('baseurl', 'tool_opensesame'),
+            get_string('baseurldesc', 'tool_opensesame'), 'https://api.delivery.opensesame.com', PARAM_URL));
 
     $settings->add(new admin_setting_configpasswordunmask('tool_opensesame/customerintegrationid',
             get_string('customerintegrationid', 'tool_opensesame'),
@@ -55,13 +54,11 @@ if ($hassiteconfig) {
 
     // Add scorm type select.
     // Types allowed.
-
     require_once($CFG->dirroot . '/mod/scorm/lib.php');
-    $options = array(
-
+    $options = [
             SCORM_TYPE_LOCAL => new lang_string('onlylocal', 'tool_opensesame'),
             SCORM_TYPE_AICCURL => new lang_string('onlyaiccurl', 'tool_opensesame')
-    );
+    ];
     $name = get_string('allowedtypes', 'tool_opensesame');
     $desc = get_string('allowedtypes_desc', 'tool_opensesame');
     $default = SCORM_TYPE_LOCAL;
@@ -85,4 +82,3 @@ if ($hassiteconfig) {
     // Add to the admin settings for opensesameintegration.
     $ADMIN->add('opensesameintegration', $settings);
 }
-
