+++ version.php
@@ -17,10 +17,13 @@
  * Version details.
  *
  * @package    tool_opensesame
- * @copyright  2023 Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @copyright  2023 Moodle
+ * @author     Felicia Wilkes <felicia.wilkes@moodle.com>
+ * @author     David Castro <david.castro@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 defined('MOODLE_INTERNAL') || die();
-$plugin->version = 2023082902; // The current plugin version (Date: YYYYMMDDXX).
-$plugin->requires = 2020061500; // Requires this Moodle version. 3.9.
+$plugin->version   = 2023082907; // The current plugin version (Date: YYYYMMDDXX).
+$plugin->requires  = 2020061500; // Requires this Moodle version. 3.9.
 $plugin->component = 'tool_opensesame'; // Full name of the plugin (used for diagnostics).
+$plugin->release   = '1.0.0';
